refactor: move AnalyticsUtil.kt into utilities package
